update:page

+++ src/main/java/kim/kin/repository/UserInfoRepository.java
@@ -1,6 +1,7 @@
 package kim.kin.repository;
 
 import kim.kin.model.UserInfo;
+import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.QueryByExampleExecutor;
@@ -13,8 +14,9 @@ import java.util.Optional;
  * @author choky
  */
 @Repository
-//public interface UserInfoRepository extends QueryByExampleExecutor<UserInfo>, CrudRepository<UserInfo,Long> {
-public interface UserInfoRepository extends PagingAndSortingRepository<UserInfo, Integer>{
+public interface UserInfoRepository extends PagingAndSortingRepository<UserInfo, Long> {
+//public interface UserInfoRepository extends QueryByExampleExecutor<UserInfo>,QuerydslPredicateExecutor<UserInfo>{
+//public interface UserInfoRepository extends PagingAndSortingRepository<UserInfo, Long>,QuerydslPredicateExecutor<UserInfo>{
 
     /**
      * 根据用户名查找用户
